move utils to components file
