--- src/Pages/AuditLog/utils/activityTypes.js
@@ -1,14 +1,0 @@
-export const activityTypes = [
-  'SIGN_UP',
-  'LOG_IN',
-  'UPDATE_USER',
-  'DELETE_USER',
-  'PRINT_PAGE',
-  'ACCESS_DOOR',
-  'CREATE_MESSAGE',
-  'DELETE_MESSAGE',
-  'VERIFY_EMAIL',
-  'EMAIL_SENT',
-  'CHANGE_PW',
-  'RESET_PW',
-];
